First version for build
This will create a DAG to decide the build order. However, project with
no dependencies at all, isn't included in the DAG right now. Those
should be added, so they can be compiled in parallel. Also, even though
we have a build order, if we would try to compile several items from
that DAG, we could very well set a side the dependencies. I.e., this
looks OK, but wouldn't work in practice. Needs to be reworked.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>