build: add build command
This will create a DAG to decide the build order. As long as there are
no dependencies it will build with as many jobs/cores available. The
building works by each and every project creating a "telios.yml" file in
the root of their own directory which has the commands that are needed
to configure and build the project. If the project doesn't have such a
file, it's possible to put a "telios.yml" file in the
<workspace>/override directory instead. However, when doing so it has to
be renamed to the same name as the git it belongs. So, if we work with
u-boot.git for example, then to override that we create (or put)
u-boot.yml file in the <workspace>/override folder.

Another feature in this patch is that it's possible to log stdout and
stderr to separate files, that will get prefixed by the name of the git
it's trying to build.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>